Extract code into helper type
